pyproject.toml
[build-system]
requires = [
    "setuptools>=63.0",
    "trove-classifiers",
    "wheel",
]

build-backend = "setuptools.build_meta"

[project]
name = "smontanaro.net"
version = "0.1.0"
description = "Skip's website"
authors = [
    { name = "Skip Montanaro", email = "skip.montanaro@gmail.com" },
]
requires-python = ">=3.10"

dependencies = [
    "arrow>=1.2.2",
    "certifi>=2022.6.15",
    "charset-normalizer>=2.1.0",
    "click>=8.1.3",
    "Flask>=2.1.3",
    "Flask-WTF>=1.0.1",
    "gunicorn>=20.1.0",
    "idna>=3.3",
    "isort>=5.10.1",
    "itsdangerous>=2.1.2",
    "Jinja2>=3.1.2",
    "lazy-object-proxy>=1.7.1",
    "MarkupSafe>=2.1.1",
    "platformdirs>=2.5.2",
    "python>=3.10",
    "python-dateutil>=2.8.2",
    "requests>=2.28.1",
    "six>=1.16.0",
    "sprdpl",
    "toml>=0.10.2",
    "urllib3>=1.26.10",
    "Werkzeug>=2.1.2",
    "wrapt>=1.14.1",
    "WTForms>=3.0.1",
]

[tool.setuptools.packages.find]
include = ["smontanaro"]

[project.urls]
"Homepage" = "https://www.smontanaro.net/"
